requirements.txt
opencv-python
imutils
numpy
Appium-Python-Client>=3.*
selenium>=4.*
Pillow
PyAutoGUI
